+++ second_semester/Web-application-database/src/main/java/data/ProductType.java
@@ -1,17 +1,17 @@
 package data;
 
 public class ProductType {
-  private final Long id;
+  private final int id;
   private String name;
   private String description;
 
-  public ProductType(Long id, String name, String description) {
+  public ProductType(int id, String name, String description) {
     this.id = id;
     this.name = name;
     this.description = description;
   }
 
-  public Long getId() {
+  public int getId() {
     return id;
   }
